+++ src/main/java/pro/sky/java/course211/homeworkSpringBucket/HomeWorkSpringBucketApplication.java
@@ -1,4 +1,4 @@
-package pro.sky.java.course211.homework_spring_bucket;
+package pro.sky.java.course211.homeworkSpringBucket;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
