+++ src/main/java/pro/sky/java/course211/homeworkSpringBucket/configuration/ConfigurationClass.java
@@ -1,14 +1,14 @@
-package pro.sky.java.course211.homework_spring_bucket.Сonfiguration;
+package pro.sky.java.course211.homeworkSpringBucket.configuration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.context.annotation.SessionScope;
-import pro.sky.java.course211.homework_spring_bucket.Repository.Bucket;
+import pro.sky.java.course211.homeworkSpringBucket.repository.Bucket;
 
 @Configuration
 public class ConfigurationClass {
     @Bean
     @SessionScope
-    public Bucket createBucket(){
+    public Bucket bucket(){
         return new Bucket();
     }
 }
