+++ src/test/java/pro/sky/java/course211/homeworkSpringBucket/HomeWorkSpringBucketApplicationTests.java
@@ -1,4 +1,4 @@
-package pro.sky.java.course211.homework_spring_bucket;
+package pro.sky.java.course211.homeworkSpringBucket;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
